32->64 bit mode skeleton, nothing yet works, for sure
svn:sync-xref-src-repo-rev: r39221

+++ src/recompiler_new/VBoxRecompiler.c
@@ -1,4 +1,4 @@
-/* $Id: VBoxRecompiler.c 13968 2008-11-07 15:48:52Z noreply@oracle.com $ */
+/* $Id: VBoxRecompiler.c 14134 2008-11-12 17:22:40Z noreply@oracle.com $ */
 /** @file
  * VBox Recompiler - QEMU.
  */
@@ -1908,7 +1908,7 @@ REMR3DECL(int)  REMR3State(PVM pVM)
     else
     {
         /* In 'normal' raw mode we don't have access to the hidden selector registers. */
-        if (pVM->rem.s.Env.segs[R_SS].selector != (uint16_t)pCtx->ss)
+        if (pVM->rem.s.Env.segs[R_SS].selector != pCtx->ss)
         {
             Log2(("REMR3State: SS changed from %04x to %04x!\n", pVM->rem.s.Env.segs[R_SS].selector, pCtx->ss));
